refactor: extract strategy form component
